+++ src/singletons/Emotes.hpp
@@ -14,7 +14,15 @@ namespace chatterino {
 class Settings;
 class Paths;
 
-class Emotes final : public Singleton
+class IEmotes
+{
+public:
+    virtual ~IEmotes() = default;
+
+    virtual ITwitchEmotes *getTwitchEmotes() = 0;
+};
+
+class Emotes final : public IEmotes, public Singleton
 {
 public:
     Emotes();
@@ -22,6 +30,11 @@ public:
     virtual void initialize(Settings &settings, Paths &paths) override;
 
     bool isIgnoredEmote(const QString &emote);
+
+    ITwitchEmotes *getTwitchEmotes() final
+    {
+        return &this->twitch;
+    }
 
     TwitchEmotes twitch;
     Emojis emojis;