+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unversioned
 
-- Major: Added support for Twitch's Chat Replies. [Wiki Page](https://wiki.chatterino.com/Features/#message-replies) (#3722, #3989, #4041, #4047, #4055, #4067)
+- Major: Added support for Twitch's Chat Replies. [Wiki Page](https://wiki.chatterino.com/Features/#message-replies) (#3722, #3989, #4041, #4047, #4055, #4067, #4077, #3905)
 - Major: Added multi-channel searching to search dialog via keyboard shortcut. (Ctrl+Shift+F by default) (#3694, #3875)
 - Major: Added support for emotes and badges from [7TV](https://7tv.app). [Wiki Page](https://wiki.chatterino.com/Third_party_services/#7tv) (#4002, #4062)
 - Major: Added support for Right-to-Left Languages (#3958)
@@ -21,6 +21,8 @@
 - Minor: Added `is:first-msg` search option. (#3700)
 - Minor: Added `is:elevated-msg` search option. (#4018)
 - Minor: Added `is:cheer-msg` search option. (#4069)
+- Minor: Added `is:redemption` search option. (#4118)
+- Minor: Added `is:reply` search option. (#4119)
 - Minor: Added `subtier:` search option (e.g. `subtier:3` to find Tier 3 subs). (#4013)
 - Minor: Added `badge:` search option (e.g. `badge:mod` to users with the moderator badge). (#4013)
 - Minor: Added AutoMod message flag filter. (#3938)
@@ -68,8 +70,11 @@
 - Minor: Migrated /vips to Helix API. Chat command will continue to be used until February 11th 2023. (#4053)
 - Minor: Migrated /uniquechat and /r9kbeta to Helix API. (#4057)
 - Minor: Migrated /uniquechatoff and /r9kbetaoff to Helix API. (#4057)
+- Minor: Migrated /commercial to Helix API. (#4094)
 - Minor: Added stream titles to windows live toast notifications. (#1297)
 - Minor: Make menus and placeholders display appropriate custom key combos. (#4045)
+- Minor: Migrated /chatters to Helix API. (#4088, #4097, #4114)
+- Minor: Migrated /mods to Helix API. (#4103)
 - Minor: Add settings tooltips. (#3437)
 - Bugfix: Connection to Twitch PubSub now recovers more reliably. (#3643, #3716)
 - Bugfix: Fixed `Smooth scrolling on new messages` setting sometimes hiding messages. (#4028)